BREAKING CHANGE: Support <section>less editing
