+++ profiles/w3c.js
@@ -32,11 +32,12 @@ const modules = [
   import("../src/core/link-to-dfn.js"),
   import("../src/core/xref.js"),
   import("../src/core/data-cite.js"),
-  import("../src/core/webidl-index.js"),
   import("../src/core/render-biblio.js"),
   import("../src/core/dfn-index.js"),
   import("../src/core/contrib.js"),
+  import("../src/core/sections.js"),
   import("../src/core/fix-headers.js"),
+  import("../src/core/webidl-index.js"),
   import("../src/core/structure.js"),
   import("../src/core/informative.js"),
   import("../src/core/id-headers.js"),
