+++ src/core/markdown.js
@@ -29,17 +29,11 @@
  *
  * will be transformed into:
  *
- *     <section>
  *       <h2>Title</h2>
- *       <section>
- *         <h3>Subtitle</h3>
- *         <p>Here's some text.</p>
- *       </section>
- *       <section>
- *         <h3>Another subtitle</h3>
- *         <p>More text.</p>
- *       </section>
- *     </section>
+ *       <h3>Subtitle</h3>
+ *       <p>Here's some text.</p>
+ *       <h3>Another subtitle</h3>
+ *       <p>More text.</p>
  *
  * The whitespace of pre elements are left alone.
  */
@@ -182,118 +176,6 @@ function workaroundBlockLevelMarkdown(element, selector) {
   }
 }
 
-class Builder {
-  constructor(doc) {
-    this.doc = doc;
-    this.root = doc.createDocumentFragment();
-    this.stack = [this.root];
-    this.current = this.root;
-  }
-  findPosition(header) {
-    return parseInt(header.tagName.charAt(1), 10);
-  }
-  findParent(position) {
-    let parent;
-    while (position > 0) {
-      position--;
-      parent = this.stack[position];
-      if (parent) return parent;
-    }
-  }
-  findHeader({ firstChild: node }) {
-    while (node) {
-      if (/H[1-6]/.test(node.tagName)) {
-        return node;
-      }
-      node = node.nextSibling;
-    }
-    return null;
-  }
-
-  addHeader(header) {
-    const section = this.doc.createElement("section");
-    const position = this.findPosition(header);
-
-    section.appendChild(header);
-    this.findParent(position).appendChild(section);
-    this.stack[position] = section;
-    this.stack.length = position + 1;
-    this.current = section;
-  }
-
-  addSection(node, process) {
-    const header = this.findHeader(node);
-    const position = header ? this.findPosition(header) : 1;
-    const parent = this.findParent(position);
-
-    if (header) {
-      node.removeChild(header);
-    }
-
-    node.appendChild(process(node));
-
-    if (header) {
-      node.prepend(header);
-    }
-
-    parent.appendChild(node);
-    this.current = parent;
-  }
-
-  addElement(node) {
-    this.current.appendChild(node);
-  }
-}
-
-function structure(fragment, doc) {
-  function process(root) {
-    const stack = new Builder(doc);
-    while (root.firstChild) {
-      const node = root.firstChild;
-      if (node.nodeType !== Node.ELEMENT_NODE) {
-        root.removeChild(node);
-        continue;
-      }
-      switch (node.localName) {
-        case "h1":
-        case "h2":
-        case "h3":
-        case "h4":
-        case "h5":
-        case "h6":
-          stack.addHeader(node);
-          break;
-        case "section":
-          stack.addSection(node, process);
-          break;
-        default:
-          stack.addElement(node);
-      }
-    }
-    return stack.root;
-  }
-  return process(fragment);
-}
-
-/**
- * Re-structure DOM around elem whose markdown has been processed.
- * @param {Element} elem
- */
-export function restructure(elem) {
-  const structuredInternals = structure(elem, elem.ownerDocument);
-  if (
-    structuredInternals.firstElementChild.localName === "section" &&
-    elem.localName === "section"
-  ) {
-    const section = structuredInternals.firstElementChild;
-    section.remove();
-    elem.append(...section.childNodes);
-  } else {
-    elem.textContent = "";
-  }
-  elem.appendChild(structuredInternals);
-}
-
 /**
  * @param {Iterable<Element>} elements
  */
@@ -317,9 +199,7 @@ export function run(conf) {
   }
   // Only has markdown-format sections
   if (!isMDFormat) {
-    for (const processedElem of processMDSections(document.body)) {
-      restructure(processedElem);
-    }
+    processMDSections(document.body);
     return;
   }
   // We transplant the UI to do the markdown processing
@@ -334,9 +214,7 @@ export function run(conf) {
   convertElement(newBody);
   // Remove links where class .nolinks
   substituteWithTextNodes(newBody.querySelectorAll(".nolinks a[href]"));
-  // Restructure the document properly
-  const fragment = structure(newBody, document);
   // Frankenstein the whole thing back together
-  newBody.append(rsUI, fragment);
+  newBody.append(rsUI);
   document.body.replaceWith(newBody);
 }
