+++ tests/spec/core/mdn-annotation.html
@@ -12,7 +12,7 @@
     </h1>
     <section id="abstract">Abstract</section>
     <section id='sotd'>sotd</section>
-    <section data-dfn-for="PaymentRequest" data-link-for="PaymentRequest">
+    <section data-dfn-for="PaymentRequest">
       <h2>
         <dfn>PaymentRequest</dfn> interface
       </h2>
@@ -21,7 +21,7 @@
           Constructor
         </h2>
       </section>
-      <section data-dfn-for="PaymentRequest" data-link-for="PaymentRequest" id="dom-paymentrequest-id">
+      <section data-dfn-for="PaymentRequest" id="dom-paymentrequest-id">
         <h2>
           <dfn>id</dfn> attribute
         </h2>
